Moving document types to types
