+++ src/server/resources/models/accountModel.ts
@@ -1,19 +1,9 @@
-import {
-  Schema,
-  model,
-  Document,
-  ObjectId,
-  Types,
-} from 'mongoose';
-import { IAccount, ICard } from '../../types';
+import { Schema, model } from 'mongoose';
 import { transformMongooseObject } from '../../utils/misc';
-
-export interface IAccountDocument extends Omit<IAccount, 'id' | 'user'>, Document {
-  _id: Types.ObjectId;
-  user: ObjectId;
-}
-export interface ICardDocument extends Omit<ICard, 'id'>, Document {}
-
+import type {
+  IAccountDocument,
+  ICardDocument,
+} from '../../types';
 /**
  * Schema for the Card
  */
