+++ src/server/resources/models/categoryModel.ts
@@ -1,18 +1,6 @@
-import {
-  Schema,
-  model,
-  Document,
-  ObjectId,
-  Types,
-} from 'mongoose';
-import { ICategory } from '../../types';
+import { Schema, model } from 'mongoose';
+import type { ICategoryDocument } from '../../types';
 import { transformMongooseObject } from '../../utils/misc';
-
-export interface ICategoryDocument extends Omit<ICategory, 'id' | 'user' | 'parentCategory'>, Document {
-  _id: Types.ObjectId;
-  user: ObjectId;
-  parentCategory: ObjectId;
-}
 
 const CategorySchema = new Schema<ICategoryDocument>({
   name: { type: String, required: true },
