+++ src/server/resources/repositories/budgetRepo.ts
@@ -1,6 +1,6 @@
-import BudgetModel, { IBudgetDocument } from '../models/budgetModel';
+import BudgetModel from '../models/budgetModel';
 import Repository from './repository';
-import { IBudget, ITransaction } from '../../types';
+import type { IBudget, IBudgetDocument, ITransaction } from '../../types';
 
 export class BudgetRepo extends Repository<IBudgetDocument, IBudget> {
   constructor(budgetModel: typeof BudgetModel = BudgetModel) {
